requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.10.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.3
cryptography==46.0.5
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.131.0
flake8==7.3.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
google-genai==1.67.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==2.8
iniconfig==2.3.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
MarkupSafe==3.0.3
mccabe==0.7.0
narwhals==2.18.0
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
passlib==1.7.4
pillow==12.1.1
pluggy==1.6.0
protobuf==6.33.5
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.11.0
pyngrok==7.5.0
PyPDF2==3.0.1
pytest==9.0.2
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
reportlab==4.4.10
requests==2.32.5
rfc3986==1.5.0
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
starlette==0.52.1
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==1.25.11
uvicorn==0.41.0
watchdog==6.0.0
websockets==16.0
